Combine GraphLoader.kt with NavigationGraph
