+++ game/src/main/resources/game.properties
@@ -11,7 +11,7 @@ floorItemsPath=./data/floor-items.yml
 objectsPath=./data/objects.yml
 stairsPath=./data/stairs.yml
 musicPath=./data/music-tracks.yml
-navGraph=./data/nav-graph.yml
+navGraphPath=./data/nav-graph.yml
 areaPath=./data/areas.yml
 objectDefinitionsPath=./data/definitions/object-definition-extras.yml
 npcDefinitionsPath=./data/definitions/npc-definition-extras.yml
